// src/app/shared/models/MetaData.model.ts
export interface MetaData {
  usrNo: number;
  roleId: number;
  langCode: string;
  pageId: number;
  componentId: number;
  ipAddress: string;
  oprNo: number;
  audOprNo: number;
  audUsrNo: number;
  funcMode: string;
  orgFuncMode: string;
  confFLag: number;
  doAuditAction: string;
  userName: string;
  delOprNo: number;
  userPassword: string;
  remark: string;
  calcObject: string[];
  wfFlag: number;
  wfAction: string;
  wfRoleGroup: string;
  printFlag: number;
  deviceType: number;
}